requirements.txt
Flask
python-dotenv
SQLAlchemy
psycopg2-binary
requests
